pyproject.toml
[tool.poetry]
name = "esta"
version = "0.1.0"
description = "tarea 3 - taller de integracion"
authors = ["TREKASH <emilio.fuenzalida@uc.cl>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
faiss-cpu = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
